# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.ErrorResponse:
    properties:
      details: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  handlers.PaginatedResponse:
    properties:
      data: {}
      pagination:
        $ref: '#/definitions/handlers.Pagination'
      success:
        type: boolean
    type: object
  handlers.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  handlers.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:12306
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/ysicing/tiga
  description: |
    Multi-instance DevOps management platform with support for MinIO, MySQL, PostgreSQL, Redis, Docker, and Caddy.

    Features:
    - Instance health monitoring and metrics collection
    - Alert management with rules and events
    - Comprehensive audit logging
    - MinIO object storage management
    - Database management with SQL editor
    - Docker container operations
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Tiga DevOps Platform API
  version: "1.0"
paths:
...
  /api/v1/audit:
    get:
      description: Get a paginated list of audit logs
      parameters:
      - description: Filter by user ID (UUID)
        in: query
        name: user_id
        type: string
      - description: Filter by resource type
        in: query
        name: resource_type
        type: string
      - description: Filter by resource ID (UUID)
        in: query
        name: resource_id
        type: string
      - description: Filter by action (create, read, update, delete)
        in: query
        name: action
        type: string
      - description: Filter by status (success, failure)
        in: query
        name: status
        type: string
      - description: Start time (RFC3339)
        in: query
        name: start_time
        type: string
      - description: End time (RFC3339)
        in: query
        name: end_time
        type: string
      - description: Filter by IP address
        in: query
        name: ip_address
        type: string
      - description: Search in description, changes
        in: query
        name: search
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List audit logs
      tags:
      - audit
  /api/v1/audit/{log_id}:
    get:
      description: Get audit log details by ID
      parameters:
      - description: Log ID (UUID)
        in: path
        name: log_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get audit log
      tags:
      - audit
  /api/v1/audit/actions:
    get:
      description: Get all distinct action types in audit logs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get distinct actions
      tags:
      - audit
  /api/v1/audit/failed:
    get:
      description: Get failed audit logs
      parameters:
      - default: 100
        description: Limit results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List failed actions
      tags:
      - audit
  /api/v1/audit/recent:
    get:
      description: Get the most recent audit logs
      parameters:
      - default: 100
        description: Limit results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List recent audit logs
      tags:
      - audit
  /api/v1/audit/resource-types:
    get:
      description: Get all distinct resource types in audit logs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get distinct resource types
      tags:
      - audit
  /api/v1/audit/resources/{resource_type}/{resource_id}:
    get:
      description: Get audit logs for a specific resource
      parameters:
      - description: Resource type
        in: path
        name: resource_type
        required: true
        type: string
      - description: Resource ID (UUID)
        in: path
        name: resource_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List resource audit logs
      tags:
      - audit
  /api/v1/audit/search:
    get:
      description: Search audit logs by text query
      parameters:
      - description: Search query
        in: query
        name: query
        required: true
        type: string
      - default: 100
        description: Limit results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search audit logs
      tags:
      - audit
  /api/v1/audit/statistics:
    get:
      description: Get overall audit log statistics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get audit statistics
      tags:
      - audit
  /api/v1/audit/timeline:
    get:
      description: Get activity timeline aggregated over time buckets
      parameters:
      - description: Start time (RFC3339)
        in: query
        name: start_time
        required: true
        type: string
      - description: End time (RFC3339)
        in: query
        name: end_time
        required: true
        type: string
      - description: Aggregation interval (e.g., '1 hour')
        in: query
        name: interval
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get activity timeline
      tags:
      - audit
...
  /api/v1/users/{user_id}/audit:
    get:
      description: Get audit logs for a specific user
      parameters:
      - description: User ID (UUID)
        in: path
        name: user_id
        required: true
        type: string
      - default: 100
        description: Limit results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List user audit logs
      tags:
      - audit
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...